// packages/material-tailwind-solid/src/types/components/carousel.ts
/**
 * This file contains the types and prop-types for Carousel component.
 */

import type { AnimationOptionsWithOverrides } from "@motionone/solid";
import type { JSX, Setter } from "solid-js";

export type prevArrow = (args: {
  loop: boolean;
  handlePrev: () => void;
  activeIndex: number;
  firstIndex: boolean;
}) => JSX.Element;

export type nextArrow = (args: {
  loop: boolean;
  handleNext: () => void;
  activeIndex: number;
  lastIndex: boolean;
}) => JSX.Element;

export type navigation = (args: {
  setActiveIndex: Setter<number>;
  activeIndex: number;
  length: number;
}) => JSX.Element;

export type autoplay = boolean;
export type autoplayDelay = number;
export type transition = AnimationOptionsWithOverrides; //????
export type loop = boolean;
export type slideRef = HTMLDivElement;

// javascript prop-types
// export const propTypesClassName: any = PropTypes.string;
// export const propTypesPrevArrow: any = PropTypes.func;
// export const propTypesNextArrow: any = PropTypes.func;
// export const propTypesNavigation: any = PropTypes.func;
// export const propTypesAutoplay: any = PropTypes.bool;
// export const propTypesAutoplayDelay: any = PropTypes.number;
// export const propTypesTransition: any = PropTypes.object;
// export const propTypesLoop: any = PropTypes.bool;
// export const propTypesChildren: any = PropTypes.node.isRequired;
// export const propTypesSlideRef: any = PropTypes.oneOfType([
//   PropTypes.func,
//   PropTypes.shape({ current: PropTypes.any }),
// ]);